drop modificata

+++ gestDB/dropTable.php
@@ -1,22 +1,32 @@
 <?php
-    include("conn.php");
+    if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
 
-    try {
-        // Nome della tabella da eliminare
-        $table_name = "user3";
+        $table = GET['tabella'];
 
-        // Query per eliminare la tabella
-        $sql = "DROP TABLE $table_name";
+        if($tabella) {
+            include("conn.php");
 
-        // Esecuzione della query
-        $conn->exec($sql);
+            try {
+                // Nome della tabella da eliminare
+                $table_name = $table;
 
-        echo "\nLa tabella '$table_name' è stata eliminata con successo.";
+                // Query per eliminare la tabella
+                $sql = "DROP TABLE $table_name";
 
-    } catch(PDOException $e) {
-        echo "Errore: " . $e->getMessage();
+                // Esecuzione della query
+                $conn->exec($sql);
+
+                echo "\nLa tabella '$table_name' è stata eliminata con successo.";
+
+            } catch(PDOException $e) {
+                echo "Errore: " . $e->getMessage();
+            }
+
+            // Chiusura della connessione
+            $conn = null;
+        } else {
+            echo "\nSpecificare nome tabella da cancellare.";
+            echo "\nSTabella non cancellata.";
+        }
     }
-
-    // Chiusura della connessione
-    $conn = null;
 ?>
